+++ mission/src/views/ItemInfo.vue
@@ -4,39 +4,7 @@
     <ItemMainImg :img="itemInfo?.productImg" />
     <SellerInfo :seller="itemInfo?.seller" />
     <ProductInfo :info="itemInfo?.itemInfo" />
-    <section class="content-container" data-test="review-section">
-      <p class="title">리뷰</p>
-      <div v-if="itemInfo?.reviews && itemInfo.reviews.length > 0">
-        <div
-          class="review-wrapper box-shadow"
-          v-for="(review, i) in itemInfo.reviews"
-          :key="i"
-          data-test="review-wrapper"
-        >
-          <div>
-            <header>
-              <span class="nickname" data-test="nickname">{{ hideNickname(review.nickname) }}</span>
-              <time
-                class="review-date"
-                data-test="review-date"
-                :datetime="review?.reviewDate"
-              >
-                {{ review.reviewDate }}
-              </time>
-            </header>
-            <strong class="review-title" data-test="review-title">{{ review.title }}</strong>
-            <p class="review-content" data-test="review-content">{{ review.content }}</p>
-          </div>
-          <img
-            v-if="review.reviewImg"
-            class="review-img"
-            :src="review.reviewImg"
-            data-test="review-img"
-          />
-        </div>
-      </div>
-      <p v-else>{{ notReviewMessage }}</p>
-    </section>
+    <ReviewInfo :reviews="itemInfo?.reviews" />
   </main>
   <div class="floating-action-btn box-shadow flex-center" data-test="floating-action-btn">
     <p data-test="floating-action-btn-content">{{ floatingActionBtnText }}</p>
@@ -45,10 +13,10 @@
 </template>
 
 <script>
-import message from '@/data/message';
 import ItemMainImg from '@/components/ItemInfo/ItemMainImg.vue';
 import SellerInfo from '@/components/ItemInfo/SellerInfo.vue';
 import ProductInfo from '@/components/ItemInfo/ProductInfo.vue';
+import ReviewInfo from '@/components/ItemInfo/ReviewInfo.vue';
 
 export default {
   name: 'ItemInfoPage',
@@ -56,6 +24,7 @@ export default {
     ItemMainImg,
     SellerInfo,
     ProductInfo,
+    ReviewInfo,
   },
   data() {
     return {
@@ -85,6 +54,7 @@ export default {
         },
         reviews: [
           {
+            id: 1,
             nickname: 'abcde',
             reviewData: '2022-01-16',
             title: '신축성이 좋아요',
@@ -92,6 +62,7 @@ export default {
             reviewImg: 'https://drive.google.com/uc?export=view&id=1up6VsQnHqeZ87I3URO1Wqqe2KpkMkYHG',
           },
           {
+            id: 2,
             nickname: 'tester',
             reviewData: '2022-01-16',
             title: '봄, 가을에 입기 딱!',
@@ -101,16 +72,7 @@ export default {
       },
     };
   },
-  methods: {
-    hideNickname(nickname) {
-      const nicknameLength = nickname.length;
-      return `${nickname.slice(0, 2)}${('*').repeat(nicknameLength - 2)}`;
-    },
-  },
   computed: {
-    notReviewMessage() {
-      return message.notReview;
-    },
     floatingActionBtnText() {
       return `${this.salesPrice} 구매`;
     },
@@ -119,57 +81,8 @@ export default {
 </script>
 
 <style scoped>
-.review-wrapper {
-  width: 100%;
-  height: 120px;
-  background: #fff;
-  border-radius: 10px;
-  margin-bottom: 20px;
-  padding: 10px;
-  display: flex;
-  justify-content: space-between;
-}
-
-.review-wrapper header {
-  display: flex;
-  align-items: flex-start;
-  margin-bottom: 6px;
-}
-
-.review-wrapper header .nickname {
-  font-weight: normal;
-  font-size: 18px;
-  margin-right: 10px;
-}
-
-.review-wrapper header .review-date {
-  font-weight: normal;
-  font-size: 14px;
-  color: var(--darkgray)
-}
-
-.review-wrapper .review-title {
-  font-weight: 600;
-  font-size: 18px;
-  margin-bottom: 4px;
-}
-
-.review-wrapper .review-content {
-  font-weight: normal;
-  font-size: 16px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  word-break: break-all;
-}
-
-.review-wrapper .review-img {
-  width: 100px;
-  height: 100px;
-  object-fit: cover;
-  margin-left: 4px;
+main {
+  padding-bottom: 80px;
 }
 
 .floating-action-btn {
